Refactor Camera Controls to own class
