+++ Assets/Project/Runtime/Scenes/Playground/Playground.unity
@@ -1250,6 +1250,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
+    - target: {fileID: 8153255993169651390, guid: 91912818565f9d547adcefcda7e00222,
+        type: 3}
+      propertyPath: cinemachineFollowCamera
+      value: 
+      objectReference: {fileID: 1185646606}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
