+++ Assets/Project/Runtime/_Scripts/Player/PlayerMovement.cs
@@ -47,36 +47,6 @@ namespace Project
         [SerializeField]
         private LayerMask groundLayers;
 
-        [Header("Cinemachine settings")]
-        [SerializeField]
-        private CinemachineVirtualCamera cinemachineFollowCamera;
-
-        [Space(10)]
-        [SerializeField]
-        [Range(-180f, 180f)]
-        private float baseCameraYaw = 0.0f;
-
-        [SerializeField]
-        [Range(-90f, 90f)]
-        private float baseCameraPitch = 45.0f;
-
-        [Space(10)]
-        [SerializeField]
-        private float cameraRotateSensitivity = 100.0f;
-
-        [Space(10)]
-        [SerializeField]
-        private float cameraZoomSensitivity = 100.0f;
-
-        [SerializeField]
-        private float cameraZoomSmoothTime = 0.12f;
-
-        [SerializeField]
-        private float cameraMinZoom = -10.0f;
-
-        [SerializeField]
-        private float cameraMaxZoom = 10.0f;
-
         // Player
         private float speed;
         private float animationBlend;
@@ -86,12 +56,6 @@ namespace Project
         private float terminalVelocity = 53.0f;
         private bool grounded;
         private Vector2 knockbackVelocity = Vector2.zero;
-
-        // cinemachine
-        private float cinemachineTargetYaw;
-        private float cinemachineTargetPitch;
-        private float cinemachineTargetZoom = 0.0f;
-        private float cinemachineZoomVelocity;
 
         // Timeouts
         private float jumpTimeoutDelta;
@@ -126,9 +90,6 @@ namespace Project
             animIDJump = Animator.StringToHash("Jump");
             animIDFreeFall = Animator.StringToHash("FreeFall");
             animIDMotionSpeed = Animator.StringToHash("MotionSpeed");
-
-            cinemachineTargetYaw = baseCameraYaw;
-            cinemachineTargetPitch = baseCameraPitch;
         }
 
         private void Update()
@@ -136,12 +97,6 @@ namespace Project
             JumpAndGravity();
             CheckGrounded();
             MovePlayer();
-        }
-
-        private void LateUpdate()
-        {
-            CameraRotation();
-            CameraZoom();
         }
 
         private void CheckGrounded()
@@ -243,48 +198,6 @@ namespace Project
 
         public void AddKnockback(Vector2 direction, float force) => knockbackVelocity += direction.normalized * force;
 
-        private void CameraRotation()
-        {
-            if (input.cameraRotation.sqrMagnitude > 0.01f)
-            {
-                cinemachineTargetYaw += input.cameraRotation.x * Time.deltaTime * cameraRotateSensitivity;
-            }
-
-            // Keep angle between -180 and 180 degrees
-            cinemachineTargetYaw = cinemachineTargetYaw - Mathf.CeilToInt(cinemachineTargetYaw / 360f) * 360f;
-            if (cinemachineTargetYaw < 0) cinemachineTargetYaw += 360f;
-
-            cinemachineTargetPitch = cinemachineTargetPitch - Mathf.CeilToInt(cinemachineTargetPitch / 360f) * 360f;
-            if (cinemachineTargetPitch < 0) cinemachineTargetPitch += 360f;
-
-            cinemachineFollowCamera.transform.rotation = Quaternion.Euler(cinemachineTargetPitch, cinemachineTargetYaw, 0);
-        }
-
-        private void CameraZoom()
-        {
-            CinemachineCameraOffset cameraOffset;
-            if (cinemachineFollowCamera.TryGetComponent(out cameraOffset) == false)
-            {
-                Debug.LogError("Follow camera needs a CinemachineCameraOffset component");
-                return;
-            }
-
-            cinemachineTargetZoom += input.cameraZoom * Time.deltaTime * cameraZoomSensitivity;
-            cinemachineTargetZoom = Mathf.Clamp(cinemachineTargetZoom, cameraMinZoom, cameraMaxZoom);
-
-            float currentCameraZoom = cameraOffset.m_Offset.z;
-
-            if (currentCameraZoom < cinemachineTargetZoom - 0.01f || currentCameraZoom > cinemachineTargetZoom + 0.01f)
-            {
-                float zoom = Mathf.SmoothDamp(cameraOffset.m_Offset.z, cinemachineTargetZoom, ref cinemachineZoomVelocity, cameraZoomSmoothTime);
-                cameraOffset.m_Offset.z = zoom;
-            }
-            else
-            {
-                cameraOffset.m_Offset.z = cinemachineTargetZoom;
-            }
-        }
-
         private void OnFootstep(AnimationEvent animationEvent)
         {
             /*
@@ -321,11 +234,6 @@ namespace Project
             Gizmos.DrawSphere(
                 new Vector3(transform.position.x, transform.position.y - groundedOffset, transform.position.z),
                 groundedRadius);
-        }
-
-        private void OnValidate()
-        {
-            if (cinemachineFollowCamera != null) cinemachineFollowCamera.transform.rotation = Quaternion.Euler(baseCameraPitch, baseCameraYaw, 0);
         }
     }
 }
